+++ NEWS.md
@@ -1,7 +1,12 @@
-# bioregion 1.0.0 (development/github version)
+# bioregion 1.0.2 (development/github version)
 
 This is a list of changes made in the development/github version of the package
 between bioregion 1.0.0 (CRAN release 2023-04-15) and the next CRAN release.
+
+* Modification of the way 'bioregion.pairwise.metric' object are controlled
+
+* Allow to (not) select 'formula metrics' in 
+similarity_dissimilarity_conversion()
 
 * Allow negative values in similarity() with the Euclidean metric